refactor `pkg` module into `package` module and submodule friends
