+++ src/bldr/topology/standalone.rs
@@ -31,7 +31,7 @@ use std::io::prelude::*;
 
 use fs::SERVICE_HOME;
 use error::{BldrResult, BldrError, ErrorKind};
-use pkg::Package;
+use package::Package;
 use state_machine::StateMachine;
 use topology::{self, State, Worker};
 use config::Config;
